Fix NetworkingError to be retryable
